+++ tests/test_track.py
@@ -554,58 +554,6 @@ class TrackTest(unittest.TestCase):
         self.assertEqual(result, mock.sentinel.link)
 
 
-@mock.patch('spotify.track.lib', spec=spotify.lib)
-class LocalTrackTest(unittest.TestCase):
-
-    def setUp(self):
-        self.session = tests.create_session()
-
-    def test_create(self, lib_mock):
-        sp_track = spotify.ffi.cast('sp_track *', 42)
-        lib_mock.sp_localtrack_create.return_value = sp_track
-
-        track = spotify.LocalTrack(
-            self.session, artist='foo', title='bar', album='baz', length=210)
-
-        self.assertEqual(track._sp_track, sp_track)
-        lib_mock.sp_localtrack_create.assert_called_once_with(
-            mock.ANY, mock.ANY, mock.ANY, 210)
-        self.assertEqual(
-            spotify.ffi.string(lib_mock.sp_localtrack_create.call_args[0][0]),
-            b'foo')
-        self.assertEqual(
-            spotify.ffi.string(lib_mock.sp_localtrack_create.call_args[0][1]),
-            b'bar')
-        self.assertEqual(
-            spotify.ffi.string(lib_mock.sp_localtrack_create.call_args[0][2]),
-            b'baz')
-        self.assertEqual(
-            lib_mock.sp_localtrack_create.call_args[0][3], 210)
-
-        # Since we *created* the sp_track, we already have a refcount of 1 and
-        # shouldn't increase the refcount when wrapping this sp_track in a
-        # Track object
-        self.assertEqual(lib_mock.sp_track_add_ref.call_count, 0)
-
-    def test_create_with_defaults(self, lib_mock):
-        sp_track = spotify.ffi.cast('sp_track *', 42)
-        lib_mock.sp_localtrack_create.return_value = sp_track
-
-        track = spotify.LocalTrack(self.session)
-
-        self.assertEqual(track._sp_track, sp_track)
-        lib_mock.sp_localtrack_create.assert_called_once_with(
-            mock.ANY, mock.ANY, mock.ANY, -1)
-        self.assertEqual(
-            lib_mock.sp_localtrack_create.call_args[0][0], spotify.ffi.NULL)
-        self.assertEqual(
-            lib_mock.sp_localtrack_create.call_args[0][1], spotify.ffi.NULL)
-        self.assertEqual(
-            lib_mock.sp_localtrack_create.call_args[0][2], spotify.ffi.NULL)
-        self.assertEqual(
-            lib_mock.sp_localtrack_create.call_args[0][3], -1)
-
-
 class TrackAvailability(unittest.TestCase):
 
     def test_has_constants(self):